backend/services/04_memory/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pinecone-client==5.0.1
httpx==0.27.2
pydantic==2.9.2
pydantic-settings==2.6.0
python-multipart==0.0.12